backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
drf-yasg==1.21.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
Pillow==10.1.0
dj-database-url==2.1.0
python-jose==3.3.0
python-jose[cryptography]==3.3.0
django-filter==23.5
django-allauth==0.58.2
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.1.0
drf-nested-routers==0.93.4
whitenoise==6.6.0
gunicorn==21.2.0
stripe==8.10.0
requests==2.32.3
